+++ epilepsy12/admin.py
@@ -5,6 +5,8 @@ from django.contrib import admin
 from django.contrib.auth.admin import UserAdmin
 from simple_history.admin import SimpleHistoryAdmin
 
+import pandas as pd
+
 # Register your models here.
 from .models import *
 from .organisational_audit import export_submission_period_as_csv
@@ -129,9 +131,13 @@ class OrganisationalAuditSubmissionPeriodAdmin(SimpleHistoryAdmin):
             submission_period = queryset.first()
             
             filename = f"e12-org-audit-{submission_period.year}.csv"
-            data = export_submission_period_as_csv(submission_period)
+            
+            rows = export_submission_period_as_csv(submission_period)
+            data = pd.DataFrame.from_dict(rows).to_csv(index=False)
 
-            response = HttpResponse("", content_type="text/csv")
+            print(f"!! {data}")
+
+            response = HttpResponse(data, content_type="text/csv")
             response['Content-Disposition'] = f"attachment; filename={filename}"
 
             return response
